fix: Fix stargate fee display issue
